services/mosquitto/client/requirements.txt
paho-mqtt>=1.6.1
clickhouse-driver>=0.2.4
pyjwt>=2.3.0
requests
clickhouse-connect
